6주차 - 게임좌표계를 실수->정수로 변경, 객체마다 타이머로 업데이트 주기 관리

+++ SnakeGame/SnakeGame/RandomSpeedObj.h
@@ -13,8 +13,8 @@ public:
     virtual void    Render() override;
 
 private:
-    float           m_XSpeed;
-    float           m_YSpeed;
+    int             m_XSpeed;
+    int             m_YSpeed;
 
     // 현재 이동방향을 오른쪽/왼쪽 구분, 위/아래 구분 하기 위한 변수이다.
     bool            m_IsRight;
