+++ SnakeGame/SnakeGame/SnakeBody.h
@@ -10,14 +10,14 @@ public:
     virtual void        Render() override;
 
 public:
-    float               GetSpeed() const { return m_Speed; }
-    void                SetSpeed(float _speed) { m_Speed = _speed; }
+    int                 GetSpeed() const { return m_Speed; }
+    void                SetSpeed(int _speed) { m_Speed = _speed; }
 
     void                OnKeyPress(BYTE _key);
     void                AddTail();
 
 private:
-    float               m_Speed;
+    int                 m_Speed;
     Direction           m_Direction;
     std::list<Object*>  m_TailList;
 };
